Refactor: Move out sns aggregator types and converters
